requirements.txt
requests>=2.31
pandas>=1.5
numpy>=1.24
matplotlib>=3.7
folium>=0.14
pyarrow>=12.0
shapely>=2.0
beautifulsoup4>=4.12
gridstatus>=0.27
pyyaml>=6.0
# Database & API
sqlalchemy>=2.0
alembic>=1.13
psycopg2-binary>=2.9
fastapi>=0.110
uvicorn>=0.27
jinja2>=3.1
